# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Timesheets",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/tsapi/public/"
        }
    ],
    "paths": {
...
        "/api/teams/{team_id}/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get team projects",
                "operationId": "a661b805310b5b840dff6c45ad50729d",
                "parameters": [
                    {
                        "name": "team_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create team",
                "operationId": "9387e9a91b7fc9526aec463f2eae7e23",
                "parameters": [
                    {
                        "name": "team_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "name": "Project name"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/teams/{team_id}/projects/{project_id}": {
            "patch": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update project",
                "operationId": "95a0583994dbcf35d39b5d77a54d720c",
                "parameters": [
                    {
                        "name": "team_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "name": "Project name"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}